Make Syck compile with -Werror=incompatible-function-pointer-types
New versions of Clang in XCode on MacOS throw a hard error on some of
the misuses of function pointer types that Syck commits. Obviously the
solution to this problem is to stop using Syck, and we're working on
that, but in the meanwhile it would be good if people could `bundle
install` applications with Syck without having to add hacks to their
bundler configs to override the flags.

Most of this commit is fixing st_hash iterators which should take
st_data_t integers as arguments, rather than pointers. There are a
couple of places where the arguments to rb_block_call need to be fixed
up to use the RB_BLOCK_CALL_FUNC_ARGLIST macro too.